complete task

+++ src/app/sub-input/sub-input.component.ts
@@ -1,13 +1,14 @@
-import { Component, OnInit, Input } from '@angular/core';
-import { FormInput } from '../model/form-inut';
+import { Component, OnInit, Input } from "@angular/core";
+import { FormInput } from "../model/form-inut";
+
+import Swal, { SweetAlertType } from "sweetalert2";
 
 @Component({
-  selector: 'app-sub-input',
-  templateUrl: './sub-input.component.html',
-  styleUrls: ['./sub-input.component.scss']
+  selector: "app-sub-input",
+  templateUrl: "./sub-input.component.html",
+  styleUrls: ["./sub-input.component.scss"]
 })
 export class SubInputComponent implements OnInit {
-
   @Input()
   parentForm: FormInput;
 
@@ -20,43 +21,38 @@ export class SubInputComponent implements OnInit {
   @Input()
   indexSub: number;
 
-  constructor() { 
-  }
+  constructor() {}
 
-  ngOnInit() {
-  }
+  ngOnInit() {}
 
   private createSubInput() {
-    
     const data = {
       condition: {
         options: this.checkType(this.parentForm.type),
         value: null
       },
-      question: '',
-      type: 'text',
+      question: "",
+      type: "text",
       parentType: this.parentForm.type,
       subForm: []
     };
 
     this.parentForm.subForm.push(data);
-
+    this.renderSwal('success', 'The form has been added');
   }
 
   private checkType(type: string) {
-    
-    if(type.toLowerCase() === 'number') {
-      return ['Equals', 'Greater than', 'Less than'];
-    } 
-    return ['Equals'];
-
+    if (type.toLowerCase() === "number") {
+      return ["Equals", "Greater than", "Less than"];
+    }
+    return ["Equals"];
   }
 
   private renderForm() {
     let template: string;
     let type = this.parentForm.type.toLowerCase();
-   
-    if(type === 'text') {
+
+    if (type === "text") {
       template = '<input type="text" />';
     }
 
@@ -64,15 +60,35 @@ export class SubInputComponent implements OnInit {
   }
 
   public deleteParentInput() {
-    
-    this.deleteSubInput(this.parent, this.indexSub);
-
+    Swal.fire({
+      title: "Are you sure?",
+      text: "Do you want to delete this form?",
+      type: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonText: "No",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes!"
+    }).then(result => {
+      if (result.value) {
+        this.deleteSubInput(this.parent, this.indexSub);
+        this.renderSwal("success", "Form has been delete.");
+      } else {
+        this.renderSwal("info", "Form is still exists.");
+      }
+    });
   }
 
   public deleteSub(parent, indexSub: number) {
-    
     parent.subForm.splice(indexSub, 1);
-
   }
 
+  private renderSwal(type: SweetAlertType, title: string) {
+    Swal.fire({
+      type: type,
+      title: title,
+      showConfirmButton: false,
+      timer: 800
+    });
+  }
 }
